myrequirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
google-api-core==1.31.5
google-api-python-client==2.36.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.7.2
google-cloud-speech==1.3.2
google-cloud-texttospeech==2.9.1
google-cloud-translate==3.0.1
google-pasta==0.2.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
h5py==3.6.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.10.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
libcst==0.4.0
Markdown==3.3.6
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
proto-plus==1.19.9
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.15.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflowjs==3.13.0
termcolor==1.1.0
tf-bodypix==0.3.8
tfjs-graph-converter==1.4.2
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
